+++ chameleon/views.py
@@ -1,12 +1,12 @@
 from django.core.urlresolvers import reverse
-from django.shortcuts import render
 from django.contrib.auth.decorators import login_required
+from django.shortcuts import render
 
 @login_required
-def dashboard( request ):
+def dashboard(request):
     actions = [
-        { 'name': 'Manage your Projects', 'url': reverse( 'projects:user_projects' ) },
-        { 'name': 'Help Desk Tickets', 'url': reverse( 'djangoRT:mytickets' ) },
-        { 'name': 'Manage Your Account', 'url': reverse( 'tas:profile' ) },
+        {'name': 'Manage your Projects', 'url': reverse('projects:user_projects')},
+        {'name': 'Help Desk Tickets', 'url': reverse('djangoRT:mytickets')},
+        {'name': 'Manage Your Account', 'url': reverse('tas:profile')},
     ]
-    return render( request, 'dashboard.html', { 'actions': actions })
+    return render(request, 'dashboard.html', {'actions': actions})