+++ chameleon_token/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+class ChameleonTokenConfig(AppConfig):
+    name = 'chameleon_token'
+    verbose_name = 'Chameleon Token'