+++ user_news/cms_models.py
@@ -1,5 +1,0 @@
-from django.db import models
-from cms.models.pluginmodel import CMSPlugin
-
-class UserNewsPluginModel(CMSPlugin):
-    limit = models.IntegerField(default=5)